Change filename, replace to single quotes, update indentation
