+++ routers/notebook-router.js
@@ -1,18 +1,18 @@
-"use strict";
+'use strict';
 
-const express = require("express"),
-  router = express.Router();
+const express = require('express'),
+    router = express.Router();
 
-const notebookController = require("../controllers/NotebookController");
+const notebookController = require('../controllers/NotebookController');
 
-router.head("/notebook/:id", notebookController.exist);
+router.head('/notebook/:id', notebookController.exist);
 
-router.post("/notebook", notebookController.create);
+router.post('/notebook', notebookController.create);
 
-router.get("/notebook/:id/notes", notebookController.getNotes);
+router.get('/notebook/:id/notes', notebookController.getNotes);
 
-router.put("/notebook/:id/notes", notebookController.updateNotes);
+router.put('/notebook/:id/notes', notebookController.updateNotes);
 
-router.put("/notebook/:id/settings", notebookController.updateSettings);
+router.put('/notebook/:id/settings', notebookController.updateSettings);
 
 module.exports = router;
